gggghgh

+++ HRMS/Models/AdvanceSalary.cs
@@ -11,7 +11,7 @@ namespace HRMS.Models
     public class AdvanceSalary
     {
         public long Id { get; set; }
-        public int? EmployeeId { get; set; }
+        public long? EmployeeId { get; set; }
         public decimal? Amount { get; set; }
         public long? PayoutMonth { get; set; }
         public bool? Active { get; set; }
